Update install_packages.R
I updated the for loop to install the devel version of specified packages as suggested in #361, It should be easy to update, if we ever need to add other devel versions of packages using remotes.

This deletes the initial line that installs all the packages that haven't been installed yet. Then installs the devel version of the specified packages and installs the remaining packages.