Small rewording
Rarefaction is defined as the repeated subsampling. This is often
confused with a single subsampling (to rarefy).
Controversy has been fueled by this confusion with rarefy/rarefaction
terminology.

I believe the simpler wording here will provide better clarity, and
avoid the use of 'rarefaction' when alluding to one iteration of
subsampling.